backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.14
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
chromadb==0.5.3
pymupdf==1.24.6
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
ragas==0.1.10
numpy==1.26.4
